test(tests): move tests to a mirror tree
Move test modules out of src into tests/, one folder per layer, so a
source folder lists only source. rootDir moves to the build config;
the base config includes both trees so typecheck keeps covering the
test tree. Test imports stay plain relative paths.

Closes #37
